added prompt utility. adress pr comments
